+++ src/trlib.pyx
@@ -9,27 +9,27 @@ cdef class Callback:
     def __cinit__(self, hvfcn):
         self.hvfcn = hvfcn
 
-cdef void hvfcn_cb(void *userdata, int n, double *d, double* hv):
+cdef void hvfcn_cb(void *userdata, long n, double *d, double* hv):
     cdef object self = <object> userdata
     _d = np.asarray(<np.float64_t[:n]> d)
     _hv = np.asarray(<np.float64_t[:n]> hv)
     _hv[:] = self.hvfcn(_d)
 
-def krylov_prepare_memory(int itmax, double[::1] fwork not None):
+def krylov_prepare_memory(long itmax, double[::1] fwork not None):
     return ctrlib.trlib_krylov_prepare_memory(itmax, &fwork[0] if fwork.shape[0] > 0 else NULL)
 
-def krylov_memory_size(int itmax):
-    cdef int iwork_size, fwork_size, h_pointer
+def krylov_memory_size(long itmax):
+    cdef long iwork_size, fwork_size, h_pointer
     ctrlib.trlib_krylov_memory_size(itmax, &iwork_size, &fwork_size, &h_pointer)
     return iwork_size, fwork_size, h_pointer
 
-def krylov_min(int init, double radius, double g_dot_g, double v_dot_g, double p_dot_Hp,
-        int[::1] iwork not None, double [::1] fwork not None,
+def krylov_min(long init, double radius, double g_dot_g, double v_dot_g, double p_dot_Hp,
+        long[::1] iwork not None, double [::1] fwork not None,
         equality = False, int itmax = 500, int itmax_lanczos = 100,
-        int ctl_invariant=0,
+        long ctl_invariant=0,
         double tol_rel_i = np.finfo(np.float).eps**.5, double tol_abs_i = 0.0,
         double tol_rel_b = np.finfo(np.float).eps**.3, double tol_abs_b = 0.0,
-        double zero = np.finfo(np.float).eps, int verbose=0, refine = True,
+        double zero = np.finfo(np.float).eps, long verbose=0, refine = True,
         long[::1] timing = None, prefix=""):
     cdef long [:] timing_b
     if timing is None:
@@ -37,7 +37,7 @@ def krylov_min(int init, double radius, double g_dot_g, double v_dot_g, double p
         timing_b = ttiming
     else:
         timing_b = timing
-    cdef int ret, action, iter, ityp
+    cdef long ret, action, iter, ityp
     cdef double flt1, flt2, flt3
     eprefix = prefix.encode('UTF-8')
     cdef char* cprefix = eprefix
